Move csvService as csvUtils
The current csv service is very generitc. Hence, it belongs in the same
directory where other similar files (ods utils...) are already present.
