+++ api/lib/domain/services/campaigns/campaign-csv-result-service.js
@@ -1,5 +1,5 @@
 const moment = require('moment');
-const csvService = require('../csv-service');
+const csvUtils = require('../../../infrastructure/utils/csv-utils');
 const _ = require('lodash');
 
 const CAMPAIGN_CSV_PLACEHOLDER = 'NA';
@@ -34,13 +34,13 @@ function createCsvHeader(campaign, { competences, areas, skillNames }) {
 
 const _headerPropertyMap = [
   { headerName: 'Nom de l\'organisation', propertyName: 'organizationName' },
-  { headerName: 'ID Campagne', propertyName: 'campaignId', type: csvService.valueTypes.NUMBER },
+  { headerName: 'ID Campagne', propertyName: 'campaignId', type: csvUtils.valueTypes.NUMBER },
   { headerName: 'Nom de la campagne', propertyName: 'campaignName' },
   { headerName: 'Nom du Profil Cible', propertyName: 'targetProfileName' },
   { headerName: 'Nom du Participant', propertyName: 'userLastName' },
   { headerName: 'Prénom du Participant', propertyName: 'userFirstName' },
-  { headerName: '% de progression', propertyName: 'progression', type: csvService.valueTypes.NUMBER },
-  { headerName: 'Date de début', propertyName: 'startedAt', type: csvService.valueTypes.NUMBER },
+  { headerName: '% de progression', propertyName: 'progression', type: csvUtils.valueTypes.NUMBER },
+  { headerName: 'Date de début', propertyName: 'startedAt', type: csvUtils.valueTypes.NUMBER },
   { headerName: 'Partage (O/N)', propertyName: 'isShared' },
 ];
 
@@ -58,17 +58,17 @@ function getHeaderPropertyMap(campaign) {
 function getHeaderPropertyMapWhenShared(campaign, { competences, areas, skills, knowledgeElements }) {
   return _.flatMap([
     getHeaderPropertyMap(campaign),
-    { headerName: 'Date du partage', propertyName: 'sharedAt', type: csvService.valueTypes.NUMBER },
-    { headerName: '% maitrise de l\'ensemble des acquis du profil', propertyName: 'knowledgeElementsValidatedPercentage', type: csvService.valueTypes.NUMBER },
+    { headerName: 'Date du partage', propertyName: 'sharedAt', type: csvUtils.valueTypes.NUMBER },
+    { headerName: '% maitrise de l\'ensemble des acquis du profil', propertyName: 'knowledgeElementsValidatedPercentage', type: csvUtils.valueTypes.NUMBER },
     _.flatMap(competences, (competence) => [
-      { headerName: `% de maitrise des acquis de la compétence ${competence.name}`, value: competence.percentage, type: csvService.valueTypes.NUMBER },
-      { headerName: `Nombre d'acquis du profil cible dans la compétence ${competence.name}`, value: competence.skillsForThisCompetence.length, type: csvService.valueTypes.NUMBER },
-      { headerName: `Acquis maitrisés dans la compétence ${competence.name}`, value: competence.numberOfSkillsValidatedForThisCompetence, type: csvService.valueTypes.NUMBER },
+      { headerName: `% de maitrise des acquis de la compétence ${competence.name}`, value: competence.percentage, type: csvUtils.valueTypes.NUMBER },
+      { headerName: `Nombre d'acquis du profil cible dans la compétence ${competence.name}`, value: competence.skillsForThisCompetence.length, type: csvUtils.valueTypes.NUMBER },
+      { headerName: `Acquis maitrisés dans la compétence ${competence.name}`, value: competence.numberOfSkillsValidatedForThisCompetence, type: csvUtils.valueTypes.NUMBER },
     ]),
     _.flatMap(areas, (area) => [
-      { headerName: `% de maitrise des acquis du domaine ${area.title}`, value:_.round(area.numberSkillsValidated / area.numberSkillsTested, 2), type: csvService.valueTypes.NUMBER },
-      { headerName: `Nombre d'acquis du profil cible du domaine ${area.title}`, value: area.numberSkillsTested, type: csvService.valueTypes.NUMBER },
-      { headerName: `Acquis maitrisés du domaine ${area.title}`, value: area.numberSkillsValidated, type: csvService.valueTypes.NUMBER },
+      { headerName: `% de maitrise des acquis du domaine ${area.title}`, value:_.round(area.numberSkillsValidated / area.numberSkillsTested, 2), type: csvUtils.valueTypes.NUMBER },
+      { headerName: `Nombre d'acquis du profil cible du domaine ${area.title}`, value: area.numberSkillsTested, type: csvUtils.valueTypes.NUMBER },
+      { headerName: `Acquis maitrisés du domaine ${area.title}`, value: area.numberSkillsValidated, type: csvUtils.valueTypes.NUMBER },
     ]),
     _.flatMap(skills, (skill) => {
       const knowledgeElementForSkill = _.find(knowledgeElements, { skillId: skill.id });
